% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{colour_to_integer}
\alias{colour_to_integer}
\title{Convert a colour as a character string  into an integer value representing RGBA}
\usage{
colour_to_integer(colour)
}
\arguments{
\item{colour}{character vector of R colour names (including hex colours)}
}
\description{
Convert a colour as a character string  into an integer value representing RGBA
}
