% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocube.R
\name{darken}
\alias{darken}
\title{Cheap version of darkening a colour (but not at all nice)}
\usage{
darken(fill, amount)
}
\arguments{
\item{fill}{vector of R colours}

\item{amount}{fraction to darken by}
}
\description{
Cheap version of darkening a colour (but not at all nice)
}
