% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{deer}
\alias{deer}
\title{Sprite sheet for an animated deer}
\format{
An object of class \code{nativeRaster} with 160 rows and 160 columns.
}
\usage{
deer
}
\description{
A sprite sheet for animated characters.  This spritesheet is in the
public domain.
}
\keyword{datasets}
