% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{deer_loc}
\alias{deer_loc}
\title{Locations of 15 sprites within the \code{deer} spritesheet}
\format{
An object of class \code{list} of length 15.
}
\usage{
deer_loc
}
\description{
List of 15 numeric vectors. Each vector consistes of (x, y, width, height)
location of a sprite within the spritesheet.
}
\details{
This 'deer' spritesheet is by Calciumtrice and licensed under CC BY 3.0.  
See 'inst/LICENSE-deer.txt' for more details.
}
\keyword{datasets}
