% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{deer_sprites}
\alias{deer_sprites}
\title{List of deer native rasters}
\format{
An object of class \code{list} of length 15.
}
\usage{
deer_sprites
}
\description{
The 'deer' sprites are by Calciumtrice and licensed under CC BY 3.0.  
See 'inst/LICENSE-deer.txt' for more details. To view license information:
}
\details{
\code{cat(readLines(system.file('LICENSE-deer.txt', package = 'nara')), sep = "\n")}
}
\keyword{datasets}
