% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dino.R
\docType{data}
\name{dino}
\alias{dino}
\title{A list of 16 frames of the blue dino from Bubble Bobble in nativeRaster
image format}
\format{
An object of class \code{list} of length 16.
}
\usage{
dino
}
\description{
A list of 16 frames of the blue dino from Bubble Bobble in nativeRaster
image format
}
\keyword{datasets}
