% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{integer_to_colour}
\alias{integer_to_colour}
\title{Convert integer values to hex colours}
\usage{
integer_to_colour(ints)
}
\arguments{
\item{ints}{integer values}
}
\value{
character vector of hex colours
}
\description{
Convert integer values to hex colours
}
