% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{is_nativeraster}
\alias{is_nativeraster}
\title{Check if object is a native raster image}
\usage{
is_nativeraster(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical. TRUE if object is a native raster image
}
\description{
Check if object is a native raster image
}
\examples{
is_nativeraster(mtcars)

}
