% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{isometric_landscape}
\alias{isometric_landscape}
\title{A list of 36 isometric tiles in nativeRaster format}
\format{
An object of class \code{list} of length 36.
}
\usage{
isometric_landscape
}
\description{
These tiles are in the public domain
}
\keyword{datasets}
