% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{king}
\alias{king}
\title{Sprite sheet for some animated characters}
\format{
An object of class \code{nativeRaster} with 128 rows and 736 columns.
}
\usage{
king
}
\description{
A sprite sheet for animated characters.  This spritesheet is in the
public domain.
}
\keyword{datasets}
