% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{king_locs}
\alias{king_locs}
\title{Locations of 16 sprites within the \code{king} spritesheet}
\format{
An object of class \code{list} of length 16.
}
\usage{
king_locs
}
\description{
List of 16 numeric vectors. Each vector consistes of (x, y, width, height)
location of a sprite within the spritesheet.
}
\keyword{datasets}
