% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit2}
\alias{nr_blit2}
\title{Blit2}
\usage{
nr_blit2(nr, x, y, src, loc, respect_alpha = TRUE)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the bottom-left
where the coordinates are (1, 1).}

\item{src}{native raster to copy from}

\item{loc}{a 4 element numeric vector (or list) with the following values 
in this exact order: (x0, y0, w, h) corresponding to the 
separate arguments to \code{nr_blit()}}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\value{
\code{nativeRaster}
}
\description{
Blit2
}
\examples{
nr <- nr_new(50, 50, 'grey80')
nr_blit2(nr, x = 1, y = 1, src = deer, loc = deer_loc[1,])
plot(nr)

}
