% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit3}
\alias{nr_blit3}
\title{Blit3}
\usage{
nr_blit3(nr, x, y, src, loc_mat, loc_idx, respect_alpha = TRUE)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the top-left
where the coordinates are (0, 0).}

\item{src}{native raster to copy from}

\item{loc_mat}{N x 4 matrix of multiple sprite locations within the \code{src}
spritesheet}

\item{loc_idx}{integer vector of rows within the \code{locs} matrix to use}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Blit3
}
\examples{
nr <- nr_new(400, 400, 'grey80')
nr_blit3(nr, x = c(0, 100, 300), y = c(0, 100, 300), 
         src = deer, loc_mat = deer_loc, loc_idx = c(1, 4, 8))
plot(nr)

}
