% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit4}
\alias{nr_blit4}
\title{Blit4}
\usage{
nr_blit4(nr, x, y, src, loc_mat, idx_mat, width, height, respect_alpha = TRUE)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the bottom-left
where the coordinates are (1, 1).}

\item{src}{native raster to copy from}

\item{loc_mat}{N x 4 matrix of multiple sprite locations within the \code{src}
spritesheet}

\item{idx_mat}{matrix of indices into \code{loc_mat}}

\item{width, height}{tile width/height (constant across all tiles)}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Blit4
}
\examples{
nr <- nr_new(400, 400, 'grey80')
idx_mat <- matrix(c(1, 2, 3, 4, 5, 6), 3, 2)
nr_blit4(nr, x = 10, y = 20, 
         src = deer, loc_mat = deer_loc, idx_mat = idx_mat, 
         width = 32, height = 32)
plot(nr)

}
