% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit_bulk}
\alias{nr_blit_bulk}
\title{Multiple blit operations in a single call}
\usage{
nr_blit_bulk(dst, src, config)
}
\arguments{
\item{dst}{destination native raster}

\item{src}{list of native rasters}

\item{config}{data.frame of configuration information for each blit which 
most contain: idx, x, y, xsrc, ysrc, w, h, hjust, vjust, respect_alpha, draw}
}
\value{
None. \code{dst} modifief by-reference and returned invisibly.
}
\description{
Multiple blit operations in a single call
}
\examples{
nr <- nr_new(90, 90, 'grey60')
config <- data.frame(
  idx = c(1, 2, 3, 4),
  x = c(10, 10, 40, 40),
  y = c(10, 40, 40, 10),
  xsrc = 0L,
  ysrc = 0L,
  w = -1L,
  h = -1L,
  hjust = 0,
  vjust = 0,
  respect_alpha = TRUE,
  render = TRUE
)
nr_blit_bulk(dst = nr, src = deer_sprites, config = config)
plot(nr)
}
