% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit_grid}
\alias{nr_blit_grid}
\title{Blit4}
\usage{
nr_blit_grid(
  nr,
  x,
  y,
  src_list,
  idx_mat,
  width,
  height,
  hjust = 0,
  vjust = 0,
  respect_alpha = TRUE
)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the top-left
where the coordinates are (0, 0).}

\item{src_list}{list of native rasters}

\item{idx_mat}{integer matrix of indices into \code{src_list}}

\item{width, height}{tile width/height (constant across all tiles)}

\item{hjust, vjust}{specify horizontal and vertical justification of the 
\code{src} image.  e.g. \code{hjust = vjust = 0} the blitting
starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
to treat the centre of the \code{src_} as the blitting origin.
Default (0, 0)}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Blit4
}
\examples{
nr <- nr_new(100, 100, 'grey80')
idx_mat <- matrix(c(
  1, 2, 3,
  4, 5, 6,
  7, 8, 9
), 3, 3, byrow = TRUE)
nr_blit_grid(nr, 0, 0, src_list = deer_sprites, idx_mat = idx_mat, width = 32, height = 32)

}
