% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit_list}
\alias{nr_blit_list}
\title{Blit from a list of native rasters}
\usage{
nr_blit_list(
  nr,
  x,
  y,
  src_list,
  src_idx,
  hjust = 0,
  vjust = 0,
  respect_alpha = TRUE
)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the top-left
where the coordinates are (0, 0).}

\item{src_list}{list of native rasters}

\item{src_idx}{indices into the list of the native raster}

\item{hjust, vjust}{specify horizontal and vertical justification of the 
\code{src} image.  e.g. \code{hjust = vjust = 0} the blitting
starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
to treat the centre of the \code{src_} as the blitting origin.
Default (0, 0)}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\description{
Blit from a list of native rasters
}
\examples{
nr <- nr_new(50, 50, 'grey80')
src_list <- list(nr_new(10, 10, 'hotpink'), nr_new(10, 10, 'blue'))
nr_blit_list(nr, x = c(0, 25), y = c(0, 25), src_list = src_list, src_idx = c(1, 2))
plot(nr)
}
