% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit-rotozoom.R
\name{nr_blit_rotozoom}
\alias{nr_blit_rotozoom}
\title{Blit a source image into another native raster with rotation and scaling.}
\usage{
nr_blit_rotozoom(nr, src, x, y, angle, scale)
}
\arguments{
\item{nr}{destination}

\item{src}{src}

\item{x, y}{location to start within dst}

\item{angle}{angle in radians}

\item{scale}{zoom factor}
}
\value{
None. \code{nr} modified in-place and returned invisibly
}
\description{
Only implements nearest neighbour interpolation.
}
\examples{
nr <- nr_new(300, 200, 'grey80')
sq <- nr_new(20, 20, 'darkblue')
nr_blit_rotozoom(nr, src = sq, x = 100, y = 100, angle = pi/3, scale = 5)
plot(nr)
}
