% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_circle}
\alias{nr_circle}
\title{Draw circles on a native raster image}
\usage{
nr_circle(nr, x, y, r, fill = "black", color = NA, use_alpha = TRUE)
}
\arguments{
\item{nr}{native raster image}

\item{x, y}{coordinates of centre of circle.  [vector]}

\item{r}{radius [vector]}

\item{fill}{interior fill color [vector]}

\item{color}{outline color. Default: NA. [vector]}

\item{use_alpha}{Use alpha channel when drawing? Logical. Default: TRUE}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Draw circles on a native raster image
}
\examples{
set.seed(1)
w <- 200
h <- 150
nr <- nr_new(w, h, 'black')

N <- 10
coords <- expand.grid(x = seq(2, w, length.out = N), 
                      y = seq(2, h, length.out = N))

cols <- sample(grDevices::terrain.colors(nrow(coords)))

nr_circle(
  nr,
  x = coords$x,
  y = coords$y,
  r = w/N/1.3,
  fill = cols
)
plot(nr)
}
\seealso{
Other drawing functions:
\code{\link[=nr_polygon]{nr_polygon()}},
\code{\link[=nr_polyline]{nr_polyline()}},
\code{\link[=nr_rect]{nr_rect()}},
\code{\link[=nr_tri_mesh]{nr_tri_mesh()}}
}
\concept{drawing functions}
