% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_copy_into}
\alias{nr_copy_into}
\title{Copy the contents of one native raster image into another.}
\usage{
nr_copy_into(dst, src)
}
\arguments{
\item{src, dst}{Source and destination native raster images}
}
\value{
Invisibly return the supplied 'dst' native raster image which was been
        modified in-place
}
\description{
The source and destination native raster images must have the same dimensions.
}
\details{
If the native raster images are of different sizes or alpha blending is
required, use the \code{nr_blit()} function.
}
\examples{
nr1 <- nr_new(200, 100, 'hotpink')
nr2 <- nr_new(200, 100, 'green')
nr_copy_into(nr1, nr2)
plot(nr1)
}
