% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_crop}
\alias{nr_crop}
\alias{nr_crop2}
\title{Crop a section out of a native raster image into a new image}
\usage{
nr_crop(nr, x, y, w, h)

nr_crop2(nr, loc)
}
\arguments{
\item{nr}{native raster image}

\item{x, y, w, h}{dimensions of cropped section}

\item{loc}{dimensions of cropped section. A vector of 4 values 
i.e. \code{c(x, y, w, h)}}
}
\value{
New native raster image
}
\description{
Crop a section out of a native raster image into a new image
}
\examples{
nr <- nr_new(400, 400, 'hotpink')
nr2 <- nr_crop(nr, 0, 0, 10, 10)
dim(nr2)
plot(nr2)
}
