% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_fill}
\alias{nr_fill}
\title{Fill a \code{nativeRaster} image with the given colour}
\usage{
nr_fill(nr, colour)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\value{
The original \code{nativeRaster} modified in-place.
}
\description{
Fill a \code{nativeRaster} image with the given colour
}
\examples{
nr <- nr_new(400, 300, 'hotpink')
nr_fill(nr, 'blue')
plot(nr)
}
