% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_flipv}
\alias{nr_flipv}
\title{Flip a native raster image vertically}
\usage{
nr_flipv(nr)
}
\arguments{
\item{nr}{native raster image}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Flip a native raster image vertically
}
\examples{
nr <- deer[[1]] |> nr_copy()
plot(nr)
nr_flipv(nr)
plot(nr)
}
\seealso{
Other transformation functions:
\code{\link[=nr_fliph]{nr_fliph()}},
\code{\link[=nr_rotate]{nr_rotate()}},
\code{\link[=nr_transpose]{nr_transpose()}}
}
\concept{transformation functions}
