% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocube.R
\name{nr_isocube}
\alias{nr_isocube}
\title{Add isocubes}
\usage{
nr_isocube(nr, x, y, ysize, fill_top, fill_left, fill_right)
}
\arguments{
\item{nr}{nativeRaster}

\item{x, y}{coords}

\item{ysize}{cube face height (in pixels)}

\item{fill_top, fill_left, fill_right}{colours for top, left and right
faces respectively. Can be single colours or vectors.}
}
\description{
Add isocubes
}
