% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocube.R
\name{nr_isocube3d}
\alias{nr_isocube3d}
\title{Create a grob of isocubes}
\usage{
nr_isocube3d(
  nr,
  coords,
  fill = "grey90",
  fill2 = NULL,
  fill3 = NULL,
  light = "top-left",
  ysize,
  xo = im_width/2,
  yo = 0,
  verbose = FALSE,
  bg_fill = "white"
)
}
\arguments{
\item{nr}{\code{nativeRaster} object}

\item{coords}{data.frame of x,y,z coordinates for the cubes (integer coordinates)}

\item{fill}{fill colour for the main face of cube as specified by the \code{light}
argument.  By default, this will be the colour of the top face of the cube,
as the default \code{light = 'top-left'}.}

\item{fill2, fill3}{fill colours for secondary and tertiary faces of
cube - as specified in the \code{light} argument.
If set to NULL (the default) then cube faces will be darkened versions
of the main \code{fill} color.}

\item{light}{direction of light.  This is a two-word argument nominating the
main light direction and secondary light direction. The default value of
'top-left' indicates the light is brightest from the top, and then
from the left side of the cube - with the right side of the cube being darkest.
This argument is only used if \code{fill2} and \code{fill3} are unspecified.
Possible values are: top-left, top-right, left-top, left-right, right-top,
right-left.}

\item{ysize}{the height in pixels of each cube face}

\item{xo, yo}{the origin of the isometric coordinate system in 'pixel' coordinates.
These values should be given as integer pixel coordinates.
Be default the origin is the middle bottom of the graphics device
i.e. \code{(xo, yo) = (im_width/2, 1)}}

\item{verbose}{Be verbose? Default FALSE.}

\item{bg_fill}{background fill for raster image. Default: white}
}
\description{
Create a grob of isocubes
}
