% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_line}
\alias{nr_line}
\title{Draw lines on a \code{nativeRaster} image}
\usage{
nr_line(
  nr,
  x1,
  y1,
  x2,
  y2,
  color = "black",
  linewidth = 1,
  mode = draw_mode$respect_alpha
)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x1, y1, x2, y2}{Vectors of coordinates of endpoints of line}

\item{color}{Color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{linewidth}{Line linewidth. Default: 1.  If \code{linewidth = 1} then a 
naive version of Bresenham is used to draw the points.  If linewidth 
is greater than 1, then the line is convert to a triangle strip and
rendered as polygons.}

\item{mode}{drawing mode.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Uses Bresenham's algorithm to draw lines. No antialiasing.
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_line(nr, x1 = c(0, N-1), y1 = c(0, 0), x2 = c(N-1, 0), y2 = c(N-1, N-1), 
        color = c('red', 'black'))
plot(nr)

}
