% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_point}
\alias{nr_point}
\title{Draw points on a native raster image}
\usage{
nr_point(nr, x, y, color = "black", use_alpha = TRUE)
}
\arguments{
\item{nr}{native raster image}

\item{x, y}{Vectors of point coordinates}

\item{color}{Vector of colors}

\item{use_alpha}{Use alpha channel when drawing? Logical. Default: TRUE}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Draw points on a native raster image
}
\examples{
w <- 200
h <- 150
nr <- nr_new(w, h, 'grey80')

coords <- expand.grid(x = seq(w) - 1, y = seq(h) - 1)
cols   <- sample(rainbow(nrow(coords)))
nr_point(nr, x = coords$x, y = coords$y, color = cols) 
plot(nr)

}
