% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polygons}
\alias{nr_polygons}
\title{Draw multiple polygon on a \code{nativeRaster} image}
\usage{
nr_polygons(nr, x, y, id = NULL, fill = "black", color = NA)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{id}{integer vector used to separate coordinates into 
multiple polygons. Consecutive runs of the same \code{id}
value belong to the same polygon.  If NULL (the default) then 
all coordinates are assumed to be vertices of a single polygon.}

\item{fill}{fill color}

\item{color}{Color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw multiple polygon on a \code{nativeRaster} image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_polygon(nr, x = c(0, N-1, 0), y = c(0, 0, N-1), fill = 'blue', color = 'red')
plot(nr)

}
