% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polyline_thick}
\alias{nr_polyline_thick}
\title{Draw thick polyline on a \code{nativeRaster} image}
\usage{
nr_polyline_thick(
  nr,
  x,
  y,
  color = "black",
  thickness = 1,
  mitre_limit = 1,
  close = FALSE
)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{color}{Color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{thickness}{Default: 1}

\item{mitre_limit}{Default: 1}

\item{close}{Should the polyline be closed? I.e. should a line be drawn between
the last point and the first point?   Default: FALSE}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw thick polyline on a \code{nativeRaster} image
}
\examples{
N <- 200
nr <- nr_new(N, N, 'grey80')
nr_polyline_thick(nr, x = c(10, N-10, 10), y = c(10, 10, N-10), color = 'red', 
                  thickness = 5, mitre_limit = 3)
nr_polyline      (nr, x = c(10, N-10, 10), y = c(10, 10, N-10), color = 'black')
plot(nr, TRUE)


N <- 200
nr <- nr_new(N, N, 'grey80')
nr_polyline_thick(nr, x = c(10, N-10, N-10, 10), y = c(10, 10, N-10, N-10), 
                  color = 'red', thickness = 10, mitre_limit = 5, close = TRUE)
nr_polyline(nr, x = c(10, N-10, N-10, 10), y = c(10, 10, N-10, N-10), 
           color = 'black', close = TRUE)
plot(nr, TRUE)

}
