% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_resize_bilinear}
\alias{nr_resize_bilinear}
\title{Scale with bilinear interpolation}
\usage{
nr_resize_bilinear(nr, width, height)
}
\arguments{
\item{nr}{native raster}

\item{width, height}{new dimensions}
}
\value{
new nativeRaster
}
\description{
Scale with bilinear interpolation
}
\examples{
stretched <- nr_resize_bilinear(deer_sprites[[1]], 100, 40)
plot(stretched)
}
