% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_resize_nn}
\alias{nr_resize_nn}
\title{Scale a nativeRaster using nearest-neighbour interpolation}
\usage{
nr_resize_nn(nr, width, height)
}
\arguments{
\item{nr}{native raster}

\item{width, height}{new dimensions}
}
\value{
new nativeRaster
}
\description{
Scale a nativeRaster using nearest-neighbour interpolation
}
\examples{
stretched <- nr_resize_nn(deer_sprites[[1]], 100, 40)
plot(stretched)
}
