% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_rotate}
\alias{nr_rotate}
\title{Rotate a native raster image by 90,180,270 degrees}
\usage{
nr_rotate(nr, angle)
}
\arguments{
\item{nr}{native raster image}

\item{angle}{one of 0,90,180,270}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Rotate a native raster image by 90,180,270 degrees
}
\examples{
nr <- nr_new(20, 10, 'hotpink')
dim(nr)
nr_rotate(nr, 90)
dim(nr)
}
\seealso{
Other transformation functions:
\code{\link[=nr_fliph]{nr_fliph()}},
\code{\link[=nr_flipv]{nr_flipv()}},
\code{\link[=nr_transpose]{nr_transpose()}}
}
\concept{transformation functions}
