% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_scale}
\alias{nr_scale}
\title{Resize a native raster image using a scale factor}
\usage{
nr_scale(nr, scale, algo = "nn")
}
\arguments{
\item{nr}{native raster image}

\item{scale}{scale factor}

\item{algo}{'nn' for nearest neighbor (the default), or 'bilinear' for 
bilinear interpolation.}
}
\value{
New native raster image
}
\description{
Resize a native raster image using a scale factor
}
\examples{
big <- deer[[1]] |> nr_scale(2)
plot(big)
}
\seealso{
Other resizing functions:
\code{\link[=nr_resize]{nr_resize()}}
}
\concept{resizing functions}
