% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-resize.R
\name{nr_swizzle}
\alias{nr_swizzle}
\title{Swizzle the pixel ordering}
\usage{
nr_swizzle(
  nr,
  order = "RGBA",
  pm_alpha = FALSE,
  target_order,
  target_pm_alpha = FALSE
)
}
\arguments{
\item{nr}{native raster}

\item{order}{Order of source pixels. Native rasters are 'RGBA'. Possible values:
'RGBA', 'BGRA', 'ARGB', 'ABGR'}

\item{pm_alpha}{Is the alpha pre-multiplied? Default: FALSE matches how
native rasters are stored.}

\item{target_order}{Order of output pixels}

\item{target_pm_alpha}{should the final alpha be pre-multiplied? Default: FALSE}
}
\value{
Original raster modified in place and returned invisibly
}
\description{
Swizzle the pixel ordering
}
\examples{
deer <- deer_sprites[[1]]
plot(deer, T)
nr_swizzle(deer, target_order = 'BGRA')
}
