% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeraster.R
\name{nr_text}
\alias{nr_text}
\title{Draw text on a \code{nativeRaster} image}
\usage{
nr_text(nr, str, x, y, colour = "black", fontsize = 8L)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{str}{character string [scalar]}

\item{x, y}{coordinates of lower-left corner of text. [scalar]}

\item{colour}{Color as a character string. Either a standard R color
       or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}
\describe{
  \item{\code{0}}{Draw the integer value \emph{as-is} into the nativeRaster. This mode
        is useful when you are using an indexed palette like many old school games
        (like AnotherWorld) and do not want the value interpreted as a colour and
        alpha blended when drawing etc}
  \item{\code{1}}{ This operation performs
       a bitwise logical-or with the current value.  This is needed for
       rendering in the \code{AnotherWorld} game engine.}
  \item{\code{2 (default)}}{Regular colour handing for standard pixel handling}
  }}

\item{fontsize}{height of font in pizels.  Only valid values are 8, 12 and 16.
Default: 8.}
}
\description{
The only font currently available is 'spleen' - a monospace
bitmap font from: https://github.com/fcambus/spleen
}
\details{
The 'spleen' font is licensed under BSD and the license is included in this
package as "LICENSE-spleen.txt"
}
