% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_text_basic}
\alias{nr_text_basic}
\title{Draw text on a \code{nativeRaster} image using the built-in \code{spleen} 
bitmapped font.}
\usage{
nr_text_basic(
  nr,
  x,
  y,
  str,
  color = "black",
  fontsize = 8L,
  mode = draw_mode$respect_alpha
)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{coordinates of lower-left corner of text}

\item{str}{character string}

\item{color}{Color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{fontsize}{height of font in pizels.  Only valid values are 8, 12 and 16.
Default: 8.}

\item{mode}{drawing mode.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
The only font currently available is 'spleen' - a monospace
bitmap font from: https://github.com/fcambus/spleen
}
\details{
The 'spleen' font is licensed under BSD and the license is included in this
package as "LICENSE-spleen.txt". 
To view LICENSE: 

\code{cat(readLines(system.file('LICENSE-spleen.txt', package = 'nara')), 
 sep = "\n")}
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_text_basic(nr, x = 0, y = N/2, str = "Hi!")
plot(nr)

}
