% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-color.R
\name{nr_threshold}
\alias{nr_threshold}
\title{Threshold to binary image}
\usage{
nr_threshold(nr, value = 0.5)
}
\arguments{
\item{nr}{native raster image}

\item{value}{Threshold value. Default: 0.5  (valid range [0, 1])}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Threshold to binary image
}
\examples{
nr <- nr_copy(deer[[1]])
plot(nr)

nr_threshold(nr, 0.9)
plot(nr)
}
\seealso{
Other color manipulation functions:
\code{\link[=nr_color_replace]{nr_color_replace()}},
\code{\link[=nr_desaturate]{nr_desaturate()}},
\code{\link[=nr_dither]{nr_dither()}}
}
\concept{color manipulation functions}
