% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_transpose}
\alias{nr_transpose}
\title{Transpose}
\usage{
nr_transpose(nr)
}
\arguments{
\item{nr}{native raster image}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Transpose
}
\examples{
nr <- nr_new(20, 10, 'hotpink')
dim(nr)
nr_transpose(nr)
dim(nr)
}
\seealso{
Other transformation functions:
\code{\link[=nr_fliph]{nr_fliph()}},
\code{\link[=nr_flipv]{nr_flipv()}},
\code{\link[=nr_rotate]{nr_rotate()}}
}
\concept{transformation functions}
