% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-tri.R
\name{nr_wireframe_mesh}
\alias{nr_wireframe_mesh}
\alias{nr_wireframe_coords}
\title{Draw an outline wireframe mesh only}
\usage{
nr_wireframe_mesh(
  nr,
  vertices,
  indices,
  col,
  draw_all = FALSE,
  format = "auto"
)

nr_wireframe_coords(nr, coords, col, draw_all = FALSE, format = "auto")
}
\arguments{
\item{nr}{native raster image}

\item{vertices}{matrix of vertex coordinates. x,y in columns (wide) or 
rows (tall).  Extra coordinates are ignored i.e. if coordiantes 
given as (x, y, z) or (x, y, z, w), 
only the first two coordinates will be used for plotting.}

\item{indices}{integer matrix of index information. 3 indices per triangle - 
either in one-triangle-per-matrix_row (tall) or 
one-triangle-per-matrix-coloumn (wide)}

\item{col}{color specification. Single color or one color per tri}

\item{draw_all}{adjust all triangles to be the correct orientation}

\item{format}{format of data. i.e. 'tall', 'wide' or 'auto'. Default: 'auto'}

\item{coords}{direct coordinate data for each triangle.  Can be 'tall' or 'wide'}
}
\value{
nr
}
\description{
Draw an outline wireframe mesh only
}
