% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{nrs_to_mp4}
\alias{nrs_to_mp4}
\title{Convert a list of native rasters to an mp4 video file}
\usage{
nrs_to_mp4(nr_list, filename, verbose = FALSE, ...)
}
\arguments{
\item{nr_list}{list of native raster image}

\item{filename}{mp4 filename}

\item{verbose}{logical. default FALSE}

\item{...}{other arguments passed to \code{magick::image_write_video()}}
}
\value{
None.
}
\description{
Convert a list of native rasters to an mp4 video file
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) withAutoprint(\{ # examplesIf}
im <- magick::logo
nr1 <- magick_to_nr(im)
nr2 <- nr_copy(nr1)
nrs_to_mp4(list(nr1, nr2), "nothing.mp4")
\dontshow{\}) # examplesIf}
}
\seealso{
Other conversion functions:
\code{\link[=magick_to_nr]{magick_to_nr()}},
\code{\link[=matrix_to_nr]{matrix_to_nr()}},
\code{\link[=nr_to_raster]{nr_to_raster()}},
\code{\link[=nrs_to_gif]{nrs_to_gif()}}
}
\concept{conversion functions}
