% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{read_png}
\alias{read_png}
\title{Read a PNG with a transparent colour into a native raster}
\usage{
read_png(source, transparent = NULL)
}
\arguments{
\item{source}{Either name of the file to read from or a raw vector 
representing the PNG file content.}

\item{transparent}{List or vector of colours which should be considered transporent
R colour names, integer values, hex colours}
}
\value{
nativeRaster
}
\description{
Read a PNG with a transparent colour into a native raster
}
