% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{str_cols_to_packed_cols}
\alias{str_cols_to_packed_cols}
\title{Convert colours (R colours and hex colours) into packed colours 
(integer values containing RGBA bytes)}
\usage{
str_cols_to_packed_cols(colours)
}
\arguments{
\item{colours}{character vector of R colour names and hex colours e.g. 
\code{c('red', 'white', NA, 'transparent', '#12345678')}}
}
\value{
Integer vector. Each integer value contains a packed colour i.e. RGBA bytes.
}
\description{
Convert colours (R colours and hex colours) into packed colours 
(integer values containing RGBA bytes)
}
\examples{
str_cols_to_packed_cols(c('red', 'white', 'blue', NA, 'transparent'))

}
