% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{tileset}
\alias{tileset}
\title{A list of nativeraster tiles used for constructing a side-scrolling game.}
\format{
An object of class \code{list} of length 44.
}
\usage{
tileset
}
\description{
The tiles in this list are of a common base size and can be conbined to 
construct various scenes.
}
\keyword{datasets}
