% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{tileset_config}
\alias{tileset_config}
\title{A data.frame of configuration information which can be used with 
\code{nr_blit_bulk()} and the \code{tileset} data.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51 rows and 6 columns.
}
\usage{
tileset_config
}
\description{
A data.frame of configuration information which can be used with 
\code{nr_blit_bulk()} and the \code{tileset} data.
}
\keyword{datasets}
