% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{transparent}
\alias{transparent}
\title{Transparent White}
\format{
An object of class \code{integer} of length 1.
}
\usage{
transparent
}
\description{
Transparent White
}
\keyword{datasets}
