% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{deer}
\alias{deer}
\title{Sprite sheet for an animated deer}
\format{
An object of class \code{nativeRaster} with 160 rows and 160 columns.
}
\usage{
deer
}
\description{
This 'deer' spritesheet is by Calciumtrice and licensed under CC BY 3.0.  
See 'inst/LICENSE-deer.txt' for more details. To view license information:
}
\details{
\code{cat(readLines(system.file('LICENSE-deer.txt', package = 'nara')), sep = "\n")}
}
\keyword{datasets}
