% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nativeraster}
\alias{is_nativeraster}
\title{Check if object is \code{nativeRaster}}
\usage{
is_nativeraster(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical. TRUE if object is a \code{nativeRaster}
}
\description{
Check if object is \code{nativeRaster}
}
\examples{
is_nativeraster(mtcars)

}
