% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{magick_to_nr}
\alias{magick_to_nr}
\alias{nr_to_magick}
\title{Convert a 'magick' image to native raster image}
\usage{
magick_to_nr(im, dst = NULL)

nr_to_magick(nr)
}
\arguments{
\item{im}{image from the \code{magick} package}

\item{dst}{destination native raster image. If NULL (the default) a 
new native raster image will be created  If a native raster image 
is supplied here, it must have the exact dimensions to match the matrix}

\item{nr}{native raster image}
}
\value{
New native raster image
}
\description{
Convert a 'magick' image to native raster image
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) withAutoprint(\{ # examplesIf}
im <- magick::logo
nr <- magick_to_nr(im)
plot(nr)
\dontshow{\}) # examplesIf}
}
\seealso{
Other conversion functions: 
\code{\link{matrix_to_nr}()},
\code{\link{nr_to_raster}()},
\code{\link{nrs_to_gif}()},
\code{\link{nrs_to_mp4}()}

Other conversion functions: 
\code{\link{matrix_to_nr}()},
\code{\link{nr_to_raster}()},
\code{\link{nrs_to_gif}()},
\code{\link{nrs_to_mp4}()}
}
\concept{conversion functions}
