% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{magick_to_nr}
\alias{magick_to_nr}
\alias{nr_to_magick}
\title{Convert a 'magick' image to \code{nativeRaster}}
\usage{
magick_to_nr(im, dst = NULL)

nr_to_magick(nr)
}
\arguments{
\item{im}{image from the \code{magick} package}

\item{dst}{destination \code{nativeRaster} object. If NULL (the default) a 
new \code{nativeRaster} will be created  If a \code{nativeRaster} 
is supplied here, it must have the exact dimensions to match the matrix}

\item{nr}{\code{nativeRaster} object}
}
\value{
\code{nativeRaster}
}
\description{
Convert a 'magick' image to \code{nativeRaster}
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
im <- magick::logo
nr <- magick_to_nr(im)
plot(nr)
\dontshow{\}) # examplesIf}
}
