% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{matrix_to_nr}
\alias{matrix_to_nr}
\title{Matrix to \code{nativeRaster}}
\usage{
matrix_to_nr(mat, palette, fill = "transparent", min = 0, max = 1, dst = NULL)
}
\arguments{
\item{mat}{integer matrix}

\item{palette}{vector of colors.  For an integer matrix, this palette must 
contain at least as many colors as the maximum integer value in \code{mat}. 
For a numeric matrix, any length palette is allowed.}

\item{fill}{Color to be used for values < 1 when input is an integer matrix.
Default: 'transparent'.}

\item{min, max}{assumed range for the numeric data.  values from the palette
will be interpolated using this range as the extents.  An error
will occur if a value lies outside this range. Default: (0, 1)}

\item{dst}{destination \code{nativeRaster} object. If NULL (the default) a 
new \code{nativeRaster} will be created  If a \code{nativeRaster} 
is supplied here, it must have the exact dimensions to match the matrix}
}
\value{
\code{nativeRaster}
}
\description{
Matrix to \code{nativeRaster}
}
\examples{
m <- matrix(1:12, 3, 4)
palette <- str_cols_to_packed_cols(rainbow(12))
nr <- matrix_to_nr(m, palette) 
plot(nr)

}
