% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{matrix_to_nr}
\alias{matrix_to_nr}
\title{Matrix to \code{nativeRaster}}
\usage{
matrix_to_nr(mat, palette, dst = NULL)
}
\arguments{
\item{mat}{integer matrix}

\item{palette}{vector of colours.  This palette must contain at least as 
many colours as the maximum integer value in \code{mat}.}

\item{dst}{destination \code{nativeRaster} object. If NULL (the default) a 
new \code{nativeRaster} will be created  If a \code{nativeRaster} 
is supplied here, it must have the exact dimensions to match the matrix}
}
\value{
\code{nativeRaster}
}
\description{
Matrix to \code{nativeRaster}
}
\examples{
m <- matrix(1:12, 3, 4)
palette <- str_cols_to_packed_cols(rainbow(12))
nr <- matrix_to_nr(m, palette) 
plot(nr)

}
