% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nara-package.R
\docType{package}
\name{nara-package}
\alias{nara}
\alias{nara-package}
\title{nara: Native Raster Image Tools}
\description{
Native rasters are a core R image format which use a compact color representation. This color representation closely aligns with graphics device internals meaning that these images can be rendered quickly. This package provides functions to quickly create, manipulate and composite native rasters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/nara}
  \item \url{https://coolbutuseless.github.io/package/nara/index.html}
  \item Report bugs at \url{https://github.com/coolbutuseless/nara/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
