% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit.R
\name{nr_blit}
\alias{nr_blit}
\title{Copy one \code{nativeRaster} image into another at an arbitrary location.}
\usage{
nr_blit(
  nr,
  x,
  y,
  src,
  x0 = 1L,
  y0 = 1L,
  w = NULL,
  h = NULL,
  respect_alpha = FALSE
)
}
\arguments{
\item{nr}{native raster to copy into}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the bottom-left
where the coordinates are (1, 1).}

\item{src}{native raster to copy from}

\item{x0, y0}{start coordiates within src}

\item{w, h}{size within src}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE.  If FALSE, then contents will be blindly overwritten
which can be much much faster.  If the \code{src} has an 
any transparent pixels, \code{respect_alpha = TRUE} is 
probably the correct setting.}
}
\value{
\code{nativeRaster}
}
\description{
This is useful as a way of positioning sprites or icons in an image.
}
\examples{
nr <- nr_new(50, 50, 'grey80')
nr_blit(nr, x = 1, y = 1, src = deer, x0 = 1, y0 = 129, w = 32, h = 32)
plot(nr)

}
