% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blit.R
\name{nr_blit}
\alias{nr_blit}
\title{Copy one \code{nativeRaster} image into another at an arbitrary location.}
\usage{
nr_blit(nr, sprite, x, y)
}
\arguments{
\item{nr}{native raster}

\item{sprite}{native raster to copy into \code{nr}}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the bottom-left
where the coordinates are (1, 1).}
}
\description{
This is useful as a way of positioning sprites or icons in an image.
}
