% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-blit-rotozoom.R
\name{nr_blit_rotozoom}
\alias{nr_blit_rotozoom}
\title{Blit a source image into another native raster with rotation and scaling.}
\usage{
nr_blit_rotozoom(
  dst,
  src,
  x,
  y,
  angle,
  scale,
  xsrc = 0L,
  ysrc = 0L,
  w = -1L,
  h = -1L,
  hjust = 0.5,
  vjust = 0.5,
  respect_alpha = TRUE
)
}
\arguments{
\item{dst, src}{source and destination native rasters}

\item{x, y}{Where in \code{nr} to place the \code{sprite}. These values must
be vectors of the same length.  If the length is greater than 1, then
the \code{sprite} will be pasted into \code{nr} at multiple locations.
Note that the
origin of \code{nativeraster} images is the top-left
where the coordinates are (0, 0).}

\item{angle}{angle in radians}

\item{scale}{zoom factor}

\item{xsrc, ysrc}{start coordiates within src}

\item{w, h}{size within src. If size is negative, then the actual width/height of
the src is used}

\item{hjust, vjust}{specify horizontal and vertical justification of the 
\code{src} image.  e.g. \code{hjust = vjust = 0} the blitting
starts at the top-left of the image. Use \code{hjust = vjust = 0.5}
to treat the centre of the \code{src_} as the blitting origin.
Default (0, 0)}

\item{respect_alpha}{Should the alpha channel be respected when blitting?
Default: TRUE means to carefully blend pixels at each location 
using alpha values.  If FALSE, then contents at the \code{dst} will 
just be replaced with \code{src} pixels
which can be much much faster.}
}
\value{
None. \code{nr} modified in-place and returned invisibly
}
\description{
Only implements nearest neighbour interpolation.
}
\examples{
nr <- nr_new(300, 200, 'grey80')
sq <- nr_new(20, 20, 'darkblue')
nr_blit_rotozoom(nr, src = sq, x = 100, y = 100, angle = pi/3, scale = 5)
plot(nr)

nr <- nr_new(300, 200, 'grey80')
sq <- png::readPNG(system.file("img", "Rlogo.png", package="png"), native = TRUE)
nr_blit_rotozoom(nr, src = sq, x = 180, y = 120, angle = pi/6, scale = 1)
plot(nr)
}
