% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_circle}
\alias{nr_circle}
\title{Draw circles on a native raster image}
\usage{
nr_circle(
  nr,
  x,
  y,
  r,
  fill = "black",
  color = NA,
  mode = draw_mode$respect_alpha
)
}
\arguments{
\item{nr}{native raster image}

\item{x, y}{coordinates of centre of circle.  [vector]}

\item{r}{radius [vector]}

\item{fill}{interior fill color [vector]}

\item{color}{outline color. Default: NA. [vector]}

\item{mode}{drawing mode.}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Draw circles on a native raster image
}
\examples{
N <- 201
nr <- nr_new(N, N, 'grey80')
nr_circle(nr, x = N/2, y = N/2, r = c(N/3, N/4), fill = c('darkred', 'black'))
plot(nr, TRUE)
}
\seealso{
Other drawing functions: 
\code{\link{nr_polygon}()},
\code{\link{nr_polyline}()},
\code{\link{nr_rect}()}
}
\concept{drawing functions}
