% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_circle}
\alias{nr_circle}
\title{Draw circles on a \code{nativeRaster} image}
\usage{
nr_circle(nr, x, y, r, fill = "black", color = NA)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{coordinates of centre of circle.  [vector]}

\item{r}{radius [vector]}

\item{fill}{interior fill color [vector]}

\item{color}{outline color. Default: NA. [vector]}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw circles on a \code{nativeRaster} image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_circle(nr, x = N/2, y = N/2, r = c(N/4, N/2), fill = c('red', 'black'))
plot(nr)

}
