% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_circle}
\alias{nr_circle}
\title{Draw circles on a \code{nativeRaster} image}
\usage{
nr_circle(nr, x, y, r, fill = "black", colour = NA)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{x, y}{coordinates of centre of circle.  [vector]}

\item{r}{radius [vector]}

\item{fill}{interior fill colour [vector]}

\item{colour}{outline colour. Default: NA. [vector]}
}
\description{
Draw circles on a \code{nativeRaster} image
}
