% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_copy}
\alias{nr_copy}
\title{Create a new \code{nativeRaster} image and copy the dimensions and
contents from an existing image}
\usage{
nr_copy(nr)
}
\arguments{
\item{nr}{\code{nativeRaster}}
}
\value{
New \code{nativeRaster}
}
\description{
Create a new \code{nativeRaster} image and copy the dimensions and
contents from an existing image
}
\examples{
nr1 <- nr_new(200, 200, 'hotpink')
nr2 <- nr_copy(nr1)
plot(nr2)
}
\seealso{
Other image creation functions: 
\code{\link{nr_new}()}
}
\concept{image creation functions}
