% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeraster.R
\name{nr_copy_into}
\alias{nr_copy_into}
\title{Copy the contents of one \code{nativeRaster} into another.}
\usage{
nr_copy_into(nr_dst, nr_src)
}
\arguments{
\item{nr_src, nr_dst}{Source and destination \code{nativeRaster} images}
}
\description{
The source and destination \code{nativeRaster} images must have the same dimensions.
}
\details{
If the \code{nativeRaster} images are of different sizes, use the
\code{nr_blit()} function.
}
