% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_crop}
\alias{nr_crop}
\title{Crop a section out of a nativeRaster}
\usage{
nr_crop(nr, x, y, w, h)
}
\arguments{
\item{nr}{nativeRaster}

\item{x, y, w, h}{dimensions of cropped section}
}
\value{
new nativeRaster
}
\description{
Crop a section out of a nativeRaster
}
