% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-color.R
\name{nr_desaturate}
\alias{nr_desaturate}
\title{Move image colors to gray}
\usage{
nr_desaturate(nr, factor = 1)
}
\arguments{
\item{nr}{native raster image}

\item{factor}{desaturation factor. Default: 1 (fully desaturate)}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Move image colors to gray
}
\examples{
plot(deer[[1]])
nr <- deer[[1]] |> nr_copy()

nr_desaturate(nr)
plot(nr)
}
\seealso{
Other color manipulation functions: 
\code{\link{nr_color_replace}()},
\code{\link{nr_dither}()},
\code{\link{nr_threshold}()}
}
\concept{color manipulation functions}
