% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-dither.R
\name{nr_desaturate}
\alias{nr_desaturate}
\title{Convert image to gray}
\usage{
nr_desaturate(nr, factor = 1)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{factor}{desaturation factor. Default: 1 (fully desaturate)}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
A naive desaturation in RGB space
}
