% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-dither.R
\name{nr_dither}
\alias{nr_dither}
\title{Dither to binary image}
\usage{
nr_dither(nr, value = 0.5, algo = "fs")
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{value}{Threshold value. Default: 0.5  (valid range [0, 1])}

\item{algo}{Dithering algorithm. 'fs' (floyd-steinberg) or 'atkinson'. 
Default: 'fs'}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Dither to binary image
}
