% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_fill}
\alias{nr_fill}
\title{Fill a \code{nativeRaster} image with the given colour}
\usage{
nr_fill(nr, colour)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{colour}{Color as a character string. Either a standard R color
       or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}
\describe{
  \item{\code{0}}{Draw the integer value \emph{as-is} into the nativeRaster. This mode
        is useful when you are using an indexed palette like many old school games
        (like AnotherWorld) and do not want the value interpreted as a colour and
        alpha blended when drawing etc}
  \item{\code{1}}{ This operation performs
       a bitwise logical-or with the current value.  This is needed for
       rendering in the \code{AnotherWorld} game engine.}
  \item{\code{2 (default)}}{Regular colour handing for standard pixel handling}
  }}
}
\description{
Fill a \code{nativeRaster} image with the given colour
}
