% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_fliph}
\alias{nr_fliph}
\title{Flip a native raster image horizontally}
\usage{
nr_fliph(nr)
}
\arguments{
\item{nr}{native raster image}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Flip a native raster image horizontally
}
\examples{
nr <- nr_new(400, 200, 'white')
nr_rect(nr, 0, 0, 30, 15)
plot(nr)
nr_fliph(nr)
plot(nr)
}
\seealso{
Other transformation functions: 
\code{\link{nr_flipv}()},
\code{\link{nr_rotate}()},
\code{\link{nr_transpose}()}
}
\concept{transformation functions}
