% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_fliph}
\alias{nr_fliph}
\title{Flip a raster horizontally}
\usage{
nr_fliph(nr)
}
\arguments{
\item{nr}{nativeRaster}
}
\value{
original nativeRaster flipped in place
}
\description{
Flip a raster horizontally
}
