% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_fliph}
\alias{nr_fliph}
\title{Flip a \code{nativeRaster} horizontally}
\usage{
nr_fliph(nr)
}
\arguments{
\item{nr}{\code{nativeRaster}}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Flip a \code{nativeRaster} horizontally
}
\examples{
nr <- nr_new(400, 200, 'white')
nr_rect(nr, 1, 1, 30, 15)
plot(nr)
nr_fliph(nr)
plot(nr)
}
