% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_flipv}
\alias{nr_flipv}
\title{Flip a raster vertically}
\usage{
nr_flipv(nr)
}
\arguments{
\item{nr}{nativeRaster}
}
\value{
original nativeRaster flipped in place
}
\description{
Flip a raster vertically
}
