% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_flipv}
\alias{nr_flipv}
\title{Flip a \code{nativeRaster} vertically}
\usage{
nr_flipv(nr)
}
\arguments{
\item{nr}{\code{nativeRaster}}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Flip a \code{nativeRaster} vertically
}
\examples{
nr <- nr_new(400, 200, 'white')
nr_rect(nr, 0, 0, 30, 15)
plot(nr)
nr_flipv(nr)
plot(nr)

}
