% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_line}
\alias{nr_line}
\title{Draw lines on a \code{nativeRaster} image}
\usage{
nr_line(nr, x0, y0, x1, y1, colour = "black")
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x0, y0, x1, y1}{Coordinates of endpoints of line. [vector]}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\description{
Uses Bresenham's algorithm to draw lines. No antialiasing.
}
