% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_line}
\alias{nr_line}
\title{Draw lines on a \code{nativeRaster} image}
\usage{
nr_line(nr, x0, y0, x1, y1, colour = "black")
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x0, y0, x1, y1}{Vectors of coordinates of endpoints of line}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\description{
Uses Bresenham's algorithm to draw lines. No antialiasing.
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_line(nr, x0 = c(1, N), y0 = c(1, 1), x1 = c(N, 1), y1 = c(N, N), 
        colour = c('red', 'black'))
plot(nr)

}
