% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-mask.R
\name{nr_mask_begin}
\alias{nr_mask_begin}
\alias{nr_mask_end}
\title{Start/end region of masked drawing}
\usage{
nr_mask_begin(nr, mask)

nr_mask_end(nr)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{mask}{Image (cairo raster) to use as mask.  This image does not need 
to be the same size as \code{cr}}
}
\value{
Invisibly return the original \code{cairo raster} object
}
\description{
These functions are used to globally set a mask
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
logo_file <- system.file("img/logo-rainbow.png", package = 'cara', mustWork = TRUE)
logo <- nr_read_png(logo_file)
plot(logo, T)
img <- nr_new(600, 400)

nr_params_set(img, cpar(line_width = 10))

nr_mask_begin(img, mask = logo)
nr_circle(img, 300, 200, 200, fill = 'blue', color = 'black')
nr_circle(img, 350, 200, 100, fill = 'darkgreen', color = 'black')
nr_mask_end(img)

plot(img, T)
\dontshow{\}) # examplesIf}
}
