% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeraster.R
\name{nr_new}
\alias{nr_new}
\title{Create a \code{nativeRaster} image}
\usage{
nr_new(width, height, fill = "white")
}
\arguments{
\item{width, height}{Image dimensions in pixels}

\item{fill}{Background fill color as a character string. Either a standard R color
or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}}
}
\description{
A \code{nativeRaster} in R looks like an integer matrix, but is interpreted
different by graphics devices:
}
\details{
\itemize{
  \item{The data should be treated as RGBA pixels in row-major ordering}
  \item{Each 32-bit integer should be interpreted as 4-bytes - one for each
        of the R, G, B and A colour channels}
}
}
