% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_point}
\alias{nr_point}
\title{Draw points on a native raster image}
\usage{
nr_point(nr, x, y, color = "black", use_alpha = TRUE)
}
\arguments{
\item{nr}{native raster image}

\item{x, y}{Vectors of point coordinates}

\item{color}{Vector of colors}

\item{use_alpha}{Use alpha channel when drawing? Logical. Default: TRUE}
}
\value{
Invisibly return the supplied native raster image which was been
        modified in-place
}
\description{
Draw points on a native raster image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_point(nr, x = seq(N), y = seq(N), color = rainbow(N)) 
plot(nr)

}
