% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_point}
\alias{nr_point}
\title{Draw points on a \code{nativeRaster} image}
\usage{
nr_point(nr, x, y, color = "black", mode = draw_mode$respect_alpha)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{color}{Vector of colors}

\item{mode}{drawing mode.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw points on a \code{nativeRaster} image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_point(nr, x = seq(N), y = seq(N), color = rainbow(N)) 
plot(nr)

}
