% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_point}
\alias{nr_point}
\title{Draw points on a \code{nativeRaster} image}
\usage{
nr_point(nr, x, y, colour = "black")
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{x, y}{Integer vectors of point/vertex coordinates}

\item{colour}{Single colour or character vector of colours the same length
as \code{x} and \code{y}}
}
\description{
Draw points on a \code{nativeRaster} image
}
