% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polygon}
\alias{nr_polygon}
\title{Draw polygon on a \code{nativeRaster} image}
\usage{
nr_polygon(nr, x, y, fill = "black", colour = NA)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{fill}{fill colour [scalar]}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw polygon on a \code{nativeRaster} image
}
