% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeraster.R
\name{nr_polygon}
\alias{nr_polygon}
\title{Draw polygon on a \code{nativeRaster} image}
\usage{
nr_polygon(nr, x, y, fill = "black", colour = NA, op = 2L)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{x, y}{Integer vectors of point/vertex coordinates}

\item{fill}{fill colour [scalar]}

\item{colour}{Color as a character string. Either a standard R color
or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}}

\item{op}{Drawing mode. Default: 2.  Possible values: 0, 1, 2
\describe{
  \item{\code{0}}{Draw the integer value \emph{as-is} into the nativeRaster. This mode
        is useful when you are using an indexed palette like many old school games
        (like AnotherWorld) and do not want the value interpreted as a colour and
        alpha blended when drawing etc}
  \item{\code{1}}{ This operation performs
       a bitwise logical-or with the current value.  This is needed for
       rendering in the \code{AnotherWorld} game engine.}
  \item{\code{2 (default)}}{Regular colour handing for standard pixel handling}
  }}
}
\description{
Draw polygon on a \code{nativeRaster} image
}
