% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polyline}
\alias{nr_polyline}
\title{Draw polyline on a \code{nativeRaster} image}
\usage{
nr_polyline(nr, x, y, colour = "black", close = FALSE)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{close}{Should the polyline be closed? I.e. should a line be drawn between
the last point and the first point?   Default: FALSE}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw polyline on a \code{nativeRaster} image
}
