% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polyline}
\alias{nr_polyline}
\title{Draw polyline on a \code{nativeRaster} image}
\usage{
nr_polyline(nr, x, y, colour = "black", close = FALSE)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{x, y}{Integer vectors of point/vertex coordinates}

\item{colour}{Color as a character string. Either a standard R color
       or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}
\describe{
  \item{\code{0}}{Draw the integer value \emph{as-is} into the nativeRaster. This mode
        is useful when you are using an indexed palette like many old school games
        (like AnotherWorld) and do not want the value interpreted as a colour and
        alpha blended when drawing etc}
  \item{\code{1}}{ This operation performs
       a bitwise logical-or with the current value.  This is needed for
       rendering in the \code{AnotherWorld} game engine.}
  \item{\code{2 (default)}}{Regular colour handing for standard pixel handling}
  }}

\item{close}{Should the polyline be closed? I.e. should a line be drawn between
the last point and the first point?   Default: FALSE}
}
\description{
Draw polyline on a \code{nativeRaster} image
}
