% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_polyline}
\alias{nr_polyline}
\title{Draw polyline on a \code{nativeRaster} image}
\usage{
nr_polyline(nr, x, y, color = "black", close = FALSE)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{Vectors of point coordinates}

\item{color}{Color as a character string. Either a standard R color
(e.g. 'blue', 'white')
or a hex color of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{close}{Should the polyline be closed? I.e. should a line be drawn between
the last point and the first point?   Default: FALSE}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw polyline on a \code{nativeRaster} image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_polyline(nr, x = c(0, N-1, 0), y = c(0, 0, N-1), color = 'red')
plot(nr)

}
