% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_rect}
\alias{nr_rect}
\title{Draw rectangles on a \code{nativeRaster} image}
\usage{
nr_rect(nr, x, y, w, h, fill = "black", colour = NA)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{coordinates of lower left corner of rectangle.  [vector]}

\item{w, h}{width and height of rectangle. [vector]}

\item{fill}{interior fill colour [vector]}

\item{colour}{outline colour. Default: NA. [vector]}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Draw rectangles on a \code{nativeRaster} image
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_rect(nr, x = c(1, N/2 + 1), y = c(1, N/2 + 1), w = N/2, h = N/4, 
        fill = 'blue', colour = c('red', 'green'))
plot(nr)

}
