% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_replace}
\alias{nr_replace}
\title{Replace colours in a native raster}
\usage{
nr_replace(nr, old, new)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{old}{Vector of old colours}

\item{new}{Vector of replacement colours}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Replace colours in a native raster
}
\examples{
nr <- nr_new(10, 10, 'hotpink')
nr_replace(nr, 'hotpink', 'grey80')
plot(nr)

}
