% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-resize.R
\name{nr_resize}
\alias{nr_resize}
\title{Resize with interpolation}
\usage{
nr_resize(nr, width, height)
}
\arguments{
\item{nr}{native raster}

\item{width, height}{target size}
}
\value{
new nativeraster
}
\description{
Resize with interpolation
}
\examples{
stretched_deer <- nr_resize(deer_sprites[[1]], 200, 32)
plot(stretched_deer, TRUE)
}
