% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_resize}
\alias{nr_resize}
\title{Scale a nativeRaster}
\usage{
nr_resize(nr, width, height, algo = "nn")
}
\arguments{
\item{nr}{native raster}

\item{width, height}{new dimensions}

\item{algo}{'nn' for nearest neighbour (the default), or 'bilinear' for 
bilinear interpolation.}
}
\value{
New \code{nativeRaster}
}
\description{
Scale a nativeRaster
}
\examples{
stretched <- nr_resize(deer_sprites[[1]], 100, 40, algo = 'nn')
plot(stretched)
}
