% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_rotate}
\alias{nr_rotate}
\title{Rotate a native raster by 90,180,270 degrees}
\usage{
nr_rotate(nr, angle)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{angle}{one of 0,90,180,270}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Rotate a native raster by 90,180,270 degrees
}
\examples{
nr <- nr_new(20, 10, 'hotpink')
dim(nr)
nr_rotate(nr, 90)
dim(nr)

}
