% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_scale}
\alias{nr_scale}
\title{Scale the size of a \code{nativeRaster} using Nearest Neighbour resizinng}
\usage{
nr_scale(nr, scale, algo = "nn")
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{scale}{scale factor}

\item{algo}{'nn' for nearest neighbour (the default), or 'bilinear' for 
bilinear interpolation.}
}
\value{
New \code{nativeRaster}
}
\description{
Scale the size of a \code{nativeRaster} using Nearest Neighbour resizinng
}
\examples{
big <- nr_scale(deer_sprites[[1]], 2)
plot(big)
}
