% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-scale.R
\name{nr_scale}
\alias{nr_scale}
\title{Scale the size of a \code{nativeRaster} using Nearest Neighbour resizinng}
\usage{
nr_scale(nr, scale)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{scale}{scale factor}
}
\value{
New \code{nativeRaster}
}
\description{
Scale the size of a \code{nativeRaster} using Nearest Neighbour resizinng
}
\examples{
nr1 <- nr_new(200, 100, 'white')
nr2 <- nr_scale(nr1, 2)

}
