% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_scale}
\alias{nr_scale}
\title{Scale up a \code{nativeRaster} by an integer factor}
\usage{
nr_scale(nr, scale)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{scale}{integer value >= 1}
}
\value{
New \code{nativeRaster}
}
\description{
Scale up a \code{nativeRaster} by an integer factor
}
\examples{
nr1 <- nr_new(200, 100, 'white')
nr2 <- nr_scale(nr1, 2)
}
