% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeraster.R
\name{nr_text}
\alias{nr_text}
\title{Draw text on a \code{nativeRaster} image}
\usage{
nr_text(nr, str, x, y, colour = "black", fontsize = 8)
}
\arguments{
\item{nr}{\code{nativeRaster} image}

\item{str}{character string [scalar]}

\item{x, y}{coordinates of lower-left corner of text. [scalar]}

\item{colour}{Color as a character string. Either a standard R color
or a hex colour of the form \code{#rrggbbaa} or \code{#rrggbb}}

\item{fontsize}{height of font in pizels.  Only valid values are 8, 12 and 16.
Default: 8.}
}
\description{
The only font currently available is 'spleen' - a monospace
bitmap font from: https://github.com/fcambus/spleen
}
\details{
The 'spleen' font is licensed under BSD and the license is included in this
package as "LICENSE-spleen.txt"
}
