% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-draw.R
\name{nr_text}
\alias{nr_text}
\title{Draw text on a \code{nativeRaster} image}
\usage{
nr_text(nr, x, y, str, colour = "black", fontsize = 8L)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{x, y}{coordinates of lower-left corner of text}

\item{str}{character string}

\item{colour}{Color as a character string. Either a standard R colour
(e.g. 'blue', 'white')
or a hex colour of the form \code{#rrggbbaa}, \code{#rrggbb}, \code{#rgba}
or \code{#rgb}}

\item{fontsize}{height of font in pizels.  Only valid values are 8, 12 and 16.
Default: 8.}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
The only font currently available is 'spleen' - a monospace
bitmap font from: https://github.com/fcambus/spleen
}
\details{
The 'spleen' font is licensed under BSD and the license is included in this
package as "LICENSE-spleen.txt"
}
\examples{
N <- 20
nr <- nr_new(N, N, 'grey80')
nr_text(nr, x = 1, y = N/3, str = "Hi!")
plot(nr)

}
