% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-dither.R
\name{nr_threshold}
\alias{nr_threshold}
\title{Threshold to binary image}
\usage{
nr_threshold(nr, value = 0.5)
}
\arguments{
\item{nr}{\code{nativeRaster}}

\item{value}{Threshold value. Default: 0.5  (valid range [0, 1])}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Threshold to binary image
}
