% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{nr_to_raster}
\alias{nr_to_raster}
\alias{raster_to_nr}
\alias{nr_to_array}
\alias{array_to_nr}
\title{Convert native raster images to/from other R objects}
\usage{
nr_to_raster(nr)

raster_to_nr(ras, dst = NULL)

nr_to_array(nr)

array_to_nr(arr, dst = NULL)
}
\arguments{
\item{nr}{native raster image}

\item{ras}{standard R raster i.e. a character matrix of hex color values}

\item{dst}{destination native raster image. If NULL (the default) a new}

\item{arr}{3d numeric array representing R,G,B,A values with dimensions [nrow, ncol, 4] or
[nrow, ncol, 3]. Each value is in range [0,1].}
}
\value{
raster, array or native raster image
}
\description{
Convert native raster images to/from other R objects
}
\examples{
nr <- nr_new(12, 8, 'hotpink')
nr_to_raster(nr)
}
\seealso{
Other conversion functions: 
\code{\link{magick_to_nr}()},
\code{\link{matrix_to_nr}()},
\code{\link{nrs_to_gif}()},
\code{\link{nrs_to_mp4}()}
}
\concept{conversion functions}
