% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{nr_to_raster}
\alias{nr_to_raster}
\alias{raster_to_nr}
\alias{nr_to_array}
\alias{array_to_nr}
\title{Convert nativeRaster images to/from other R objects}
\usage{
nr_to_raster(nr)

raster_to_nr(ras, dst = NULL)

nr_to_array(nr)

array_to_nr(arr, dst = NULL)
}
\arguments{
\item{nr}{nativeRaster object}

\item{ras}{standard R raster i.e. a character matrix of hex colour values}

\item{dst}{destination nativeRaster. If NULL (the default) a new nativeRaster
will be created.}

\item{arr}{3d numeric array representing R,G,B,A values with dimensions [nrow, ncol, 4] or
[nrow, ncol, 3]. Each value is in range [0,1].}
}
\value{
raster, array or nativeRaster
}
\description{
Convert nativeRaster images to/from other R objects
}
