% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{nr_transpose}
\alias{nr_transpose}
\title{Transpose}
\usage{
nr_transpose(nr)
}
\arguments{
\item{nr}{\code{nativeRaster}}
}
\value{
Original \code{nativeRaster} modified in-place
}
\description{
Transpose
}
\examples{
nr <- nr_new(20, 10, 'hotpink')
dim(nr)
nr_transpose(nr)
dim(nr)

}
