% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-tri.R
\name{nr_tri_mesh}
\alias{nr_tri_mesh}
\alias{nr_tri_coords}
\title{Draw multiple triangles from mesh data}
\usage{
nr_tri_mesh(
  nr,
  vertices,
  indices,
  col,
  tri_mode = "ccw",
  mode = draw_mode$respect_alpha
)

nr_tri_coords(
  nr,
  coords,
  col,
  tri_mode = "ccw",
  mode = draw_mode$respect_alpha
)
}
\arguments{
\item{nr}{native raster image}

\item{vertices}{matrix of vertex coordinates. x,y in columns (wide) or 
rows (tall).  Extra coordinates are ignored i.e. if coordiantes 
given as (x, y, z) or (x, y, z, w), 
only the first two coordinates will be used for plotting.}

\item{indices}{integer matrix of index information. 3 indices per triangle - 
either in one-triangle-per-matrix_row (tall) or 
one-triangle-per-matrix-column (wide)}

\item{col}{color specification. Single color or one color per tri}

\item{tri_mode}{'ccw', 'cw', 'all'}

\item{mode}{drawing mode.}

\item{coords}{Numeric atrix of direct coordinate data for each triangle.
The coordinates for
each vertex are in columns. Each group of 3 columns defines one
triangle.  There may be extra rows in the 
matrix, but only the first two will be used (as x and y 
respectively)}
}
\value{
Invisibly return native raster
}
\description{
Draw multiple triangles from mesh data
}
\seealso{
Other drawing functions: 
\code{\link{nr_circle}()},
\code{\link{nr_polygon}()},
\code{\link{nr_polyline}()},
\code{\link{nr_rect}()}
}
\concept{drawing functions}
