% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-tri.R
\name{nr_tri_mesh}
\alias{nr_tri_mesh}
\alias{nr_tri_coords}
\title{Draw multiple triangles from mesh data}
\usage{
nr_tri_mesh(nr, vertices, indices, col, draw_all = FALSE)

nr_tri_coords(nr, coords, col, tri_mode = "ccw")
}
\arguments{
\item{nr}{native raster image}

\item{vertices}{matrix of vertex coordinates. x,y in columns (wide) or 
rows (tall).  Extra coordinates are ignored i.e. if coordiantes 
given as (x, y, z) or (x, y, z, w), 
only the first two coordinates will be used for plotting.}

\item{indices}{integer matrix of index information. 3 indices per triangle - 
either in one-triangle-per-matrix_row (tall) or 
one-triangle-per-matrix-column (wide)}

\item{col}{color specification. Single color or one color per tri}

\item{coords}{direct coordinate data for each triangle.  Can be 'tall' or 'wide'}

\item{tri_mode}{'ccw', 'cw', 'all'}
}
\value{
Invisibly return native raster
}
\description{
Draw multiple triangles from mesh data
}
\seealso{
Other drawing functions: 
\code{\link{nr_circle}()},
\code{\link{nr_polygon}()},
\code{\link{nr_polyline}()},
\code{\link{nr_rect}()}
}
\concept{drawing functions}
