% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{nrs_to_gif}
\alias{nrs_to_gif}
\title{Magick conversion - possibly won't be in final version of 'nara' pkg}
\usage{
nrs_to_gif(nr_list, filename, verbose = FALSE, framerate = 30, ...)
}
\arguments{
\item{nr_list}{list of native raster image}

\item{filename}{filename of gif}

\item{verbose}{logical. default FALSE}

\item{framerate}{frames per second}

\item{...}{other arguments passed to \code{magick::image_write_gif()}}
}
\value{
None.
}
\description{
Magick conversion - possibly won't be in final version of 'nara' pkg
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) withAutoprint(\{ # examplesIf}
im <- magick::logo
nr1 <- magick_to_nr(im)
nr2 <- nr_copy(nr1)
nrs_to_gif(list(nr1, nr2), "nothing.gif")
\dontshow{\}) # examplesIf}
}
\seealso{
Other conversion functions: 
\code{\link{magick_to_nr}()},
\code{\link{matrix_to_nr}()},
\code{\link{nr_to_raster}()},
\code{\link{nrs_to_mp4}()}
}
\concept{conversion functions}
