% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{nrs_to_gif}
\alias{nrs_to_gif}
\title{Magick conversion - possibly won't be in final version of 'nara' pkg}
\usage{
nrs_to_gif(nr_list, gif_name, verbose = FALSE, framerate = 30, ...)
}
\arguments{
\item{nr_list}{list of nativeRasters}

\item{gif_name}{name of mp4 file to save}

\item{verbose}{logical. default FALSE}

\item{framerate}{frames per second}

\item{...}{other arguments passed to \code{magick::image_write_gif()}}
}
\description{
Magick conversion - possibly won't be in final version of 'nara' pkg
}
