% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{nrs_to_gif}
\alias{nrs_to_gif}
\title{Magick conversion - possibly won't be in final version of 'nara' pkg}
\usage{
nrs_to_gif(nr_list, filename, verbose = FALSE, framerate = 30, ...)
}
\arguments{
\item{nr_list}{list of nativeRasters}

\item{filename}{filename of gif}

\item{verbose}{logical. default FALSE}

\item{framerate}{frames per second}

\item{...}{other arguments passed to \code{magick::image_write_gif()}}
}
\value{
None.
}
\description{
Magick conversion - possibly won't be in final version of 'nara' pkg
}
\examples{
\dontshow{if (interactive() && requireNamespace('magick', quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
im <- magick::logo
nr1 <- magick_to_nr(im)
nr2 <- nr_duplicate(nr1)
nrs_to_gif(list(nr1, nr2), "nothing.gif")
\dontshow{\}) # examplesIf}
}
