% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-magick.R
\name{nrs_to_mp4}
\alias{nrs_to_mp4}
\title{Magick conversion - possible won't be in final version of 'nara' pkg}
\usage{
nrs_to_mp4(nr_list, mp4_name, verbose = FALSE)
}
\arguments{
\item{nr_list}{list of nativeRasters}

\item{mp4_name}{name of mp4 file to save}

\item{verbose}{logical. default FALSE}
}
\description{
Magick conversion - possible won't be in final version of 'nara' pkg
}
