% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{packed_cols_to_hex_cols}
\alias{packed_cols_to_hex_cols}
\title{Convert packed colors (integer values containing RGBA bytes) to hex colors}
\usage{
packed_cols_to_hex_cols(packed_cols)
}
\arguments{
\item{packed_cols}{integer values each containing packed RGBA color information}
}
\value{
character vector of hex colors
}
\description{
Convert packed colors (integer values containing RGBA bytes) to hex colors
}
\examples{
packed_cols_to_hex_cols(c(-16776961L, -1L, -65536L, 16777215L, 16777215L))


}
