% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{packed_cols_to_hex_cols}
\alias{packed_cols_to_hex_cols}
\title{Convert packed colours (integer values containing RGBA bytes) to hex colours}
\usage{
packed_cols_to_hex_cols(packed_cols)
}
\arguments{
\item{packed_cols}{integer values each containing packed RGBA colour information}
}
\value{
character vector of hex colours
}
\description{
Convert packed colours (integer values containing RGBA bytes) to hex colours
}
