% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-core.R
\name{plot.nativeRaster}
\alias{plot.nativeRaster}
\title{Plot a native raster image}
\usage{
\method{plot}{nativeRaster}(x, y, ...)
}
\arguments{
\item{x}{native raster image}

\item{y}{any argument here will cause \code{grid::grid.newpage()} to 
be called prior to drawing the image}

\item{...}{other arguments passed to \code{grid::grid.raster()}}
}
\value{
Invisibly return the supplied native raster image
}
\description{
Plot a native raster image
}
\examples{
nr <- nr_new(200, 100, 'hotpink')
plot(nr)

}
