% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.nativeRaster}
\alias{plot.nativeRaster}
\title{Plot a native raster as an image}
\usage{
\method{plot}{nativeRaster}(x, y, ...)
}
\arguments{
\item{x}{native raster}

\item{y}{any argument here will cause the page to be called using a call
to \code{grid::grid.newpage()}}

\item{...}{other arguments passed to \code{grid::grid.raster()}}
}
\value{
None.
}
\description{
Plot a native raster as an image
}
