% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.nativeRaster}
\alias{plot.nativeRaster}
\title{Plot a \code{nativeRaster} as an image}
\usage{
\method{plot}{nativeRaster}(x, y, ...)
}
\arguments{
\item{x}{\code{nativeRaster}}

\item{y}{any argument here will cause \code{grid::grid.newpage()} to 
be called prior to drawing the \code{nativeRaster}}

\item{...}{other arguments passed to \code{grid::grid.raster()}}
}
\value{
None.
}
\description{
Plot a \code{nativeRaster} as an image
}
\examples{
nr <- nr_new(200, 100, 'hotpink')
plot(nr)

}
