% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{read_png}
\alias{read_png}
\title{Read PNG as \code{nativeRaster} converting specified colours to transparent}
\usage{
read_png(source, make_transparent = NULL)
}
\arguments{
\item{source}{Either name of the file to read from or a raw vector 
representing the PNG file content.}

\item{make_transparent}{List or vector of colours which should be considered transporent.
R colour names, hex colours, or packed colours (integers)}
}
\value{
\code{nativeRaster}
}
\description{
Read PNG as \code{nativeRaster} converting specified colours to transparent
}
\examples{
source <- system.file("img/Rlogo.png", package = "png", mustWork = TRUE)
nr <- read_png(source, make_transparent = 'white')
plot(nr)

}
