% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{str_cols_to_packed_cols}
\alias{str_cols_to_packed_cols}
\title{Convert colors (R colors and hex colors) into packed colors 
(integer values containing RGBA bytes)}
\usage{
str_cols_to_packed_cols(colors)
}
\arguments{
\item{colors}{character vector of R color names and hex colors e.g. 
\code{c('red', 'white', NA, 'transparent', '#12345678')}}
}
\value{
Integer vector. Each integer value contains a packed color i.e. RGBA bytes.
}
\description{
Convert colors (R colors and hex colors) into packed colors 
(integer values containing RGBA bytes)
}
\examples{
str_cols_to_packed_cols(c('red', 'white', 'blue', NA, 'transparent'))

}
